+++ oblate/fields/typings.py
@@ -22,25 +22,69 @@
 
 from __future__ import annotations
 
-from typing import Any, TYPE_CHECKING
 from oblate.fields.base import Field
 
-if TYPE_CHECKING:
-    from oblate.contexts import LoadContext, DumpContext
+# typing imported as t to avoid name conflict with classes
+# defined in this module
+import typing as t
+
+if t.TYPE_CHECKING:  # pragma: no cover
+    from oblate.contexts import LoadContext, DumpContext, ErrorContext
+    from oblate.exceptions import FieldError
 
 __all__ = (
     'Any',
+    'Literal',
 )
 
+_T = t.TypeVar('_T')
 
-class Any(Field[Any, Any]):
+
+class Any(Field[t.Any, t.Any]):
     """A field that accepts any arbitrary value.
 
     This field acts as a "raw field" that performs no validation on the
     given value.
     """
+    __slots__ = ()
+
     def value_load(self, value: Any, context: LoadContext) -> Any:
         return value
 
     def value_dump(self, value: Any, context: DumpContext) -> Any:
+        return value
+
+
+class Literal(Field[_T, _T]):
+    """A field that accepts only exact literal values.
+
+    This works in a similar fashion as :class:`typing.Literal`.
+
+    Parameters
+    ----------
+    *values:
+        The literal values.
+    """
+    __slots__ = ('_values', '_msg')
+
+    ERR_INVALID_VALUE = 'literal.invalid_value'
+
+    def __init__(self, *values: _T, **kwargs: t.Any) -> None:
+        self._values = values
+        self._msg = f'Value must be one of: {", ".join(repr(value) for value in self._values)}'
+        super().__init__(**kwargs)
+
+    def format_error(self, error_code: t.Any, context: ErrorContext) -> t.Union[FieldError, str]:
+        if error_code == self.ERR_INVALID_VALUE:
+            return self._msg
+
+        return super().format_error(error_code, context)  # pragma: no cover
+
+    def value_load(self, value: t.Any, context: LoadContext) -> _T:
+        if value not in self._values:
+            raise self._call_format_error(self.ERR_INVALID_VALUE, context.schema, value)
+
+        return value
+
+    def value_dump(self, value: _T, context: DumpContext) -> t.Any:
         return value