+++ docs/source/changelog.rst
@@ -12,7 +12,8 @@ New features
 ~~~~~~~~~~~~
 
 - Add :class:`validate.Range` validator.
-- Add :class:`fields.Any` as a no validation field
+- Add :class:`fields.Any` field as a no validation field.
+- Add :class:`fields.Literal` field to accept pre-defined literal values.
 - Add ``init_kwargs`` parameter in :class:`fields.Object` to support passing initialization arguments.
 
 Improvements